backend/requirements.txt
stable-ts
fastapi
uvicorn
python-multipart
openai
python-dotenv
pydantic-settings